pyproject.toml
[tool.poetry]
name = "cloud-automation"
version = "0.1.0"
description = ""
authors = ["kunal-goyal <imkunalgoyal@gmail.com>"]
readme = "README.md"
packages = [{include = "cloud_automation"}]

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-compute = "^1.15.0"
flask = "^3.0.1"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
